move pomwars-specific modules to lib
